Сливаю код

+++ example-for-graduate-work-main/src/main/java/ru/skypro/homework/dto/ExtendedAdDTO.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class ExtendedAdDTO {
-    Long adId;
+    Integer adId;
     private String userName;
     private String userSurname;
     private String userLogin;
